Implement methods for arrays fetch
Array-based Tumblr APIs can fetch maximum 20 items for each API call.
That requires that the UI would have to manage multiple calls, in order
to retrieve the full list.

This change implements a internal logic in the TumblrClient class,
that performs the whole sequence automatically, so the UI will get
the final result, without bothering with the intermediate calls.