Add single thread Executor
This change adds a single thread Executor to the TumblrClient class,
so the UI can push as many API calls as needed, and they'll be serialized
in a FCFS order.

This simplifies a lot the UI logic, as it doesn't have to care anymore
about the termination of the current call before executing a new one.