Fix infinite loop in array-based API calls
This patch implements a test for the Followers/Following API calls.
When a limit == -1 is used, the API does not terminate.

An early check is implemented, to make sure that we do not stuck executing
infinite API calls, when the number of total items is less that what the API states.